Release downloading complete
Fixed USB drive selection (when drive is not formatted)
Added parallel progress reporting
Progress reports now in MB during Download and in files during USB preparation

+++ SonmInstaller/ReleaseMetadata.fs
@@ -20,7 +20,7 @@ type Archive = {
 type Component = {
     Name: string
     Archives: Archive list
-    Size: int64
+    Size: int
 }
 
 type Version = {
